[Nexters/covid-letter-web#73] ♻️ logout 함수 MainLayout으로 이동
